Do context a little differently
I think this is more the intended approach.  It's definitely a little
confusing the way errgroup wants you to do it, but I think they do it
that way for the exact same reason you hit: to avoid storing the context
inside the errgroup.  Because of that, instead of
```
g := errgroup.WithContext(ctx)
g.Go(func(ctx context.Context) error {
    // use ctx
})
```
they designed it as
```
g, ctx := errgroup.WithContext(ctx)
g.Go(func() error {
    // use ctx
})
```

Note that closing over the context the way I do in the examples is
probably necessary in real code even if you ignore the done-checking:
you can imagine you replace the ticker with something that may want the
context more generally (say, an HTTP request).